auth_standalone/requirements_auth.txt
# Authentication dependencies for 佛學普化小助手
fastapi-users[sqlalchemy]==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver
resend==0.6.0
python-dotenv==1.0.0